move profile related files into new "profiles" directory
