+++ xbmc/GUIPassword.cpp
@@ -25,8 +25,8 @@
 #include "guilib/GUIKeyboardFactory.h"
 #include "dialogs/GUIDialogNumeric.h"
 #include "dialogs/GUIDialogOK.h"
-#include "settings/dialogs/GUIDialogLockSettings.h"
-#include "settings/dialogs/GUIDialogProfileSettings.h"
+#include "profiles/dialogs/GUIDialogLockSettings.h"
+#include "profiles/dialogs/GUIDialogProfileSettings.h"
 #include "Util.h"
 #include "settings/Settings.h"
 #include "settings/GUISettings.h"
