+++ xbmc/profiles/Makefile
@@ -1,0 +1,6 @@
+SRCS=Profile.cpp \
+
+LIB=profiles.a
+
+include ../../Makefile.include
+-include $(patsubst %.cpp,%.P,$(patsubst %.c,%.P,$(SRCS)))
