+++ xbmc/profiles/dialogs/Makefile
@@ -1,7 +1,7 @@
-SRCS=GUIDialogContentSettings.cpp \
-     GUIDialogSettings.cpp \
+SRCS=GUIDialogLockSettings.cpp \
+     GUIDialogProfileSettings.cpp \
 
-LIB=settings_dialogs.a
+LIB=profiles_dialogs.a
 
 include ../../../Makefile.include
 -include $(patsubst %.cpp,%.P,$(patsubst %.c,%.P,$(SRCS)))
