+++ xbmc/profiles/windows/GUIWindowSettingsProfile.h
