+++ xbmc/windows/GUIWindowLoginScreen.cpp
@@ -22,9 +22,10 @@
 #include "Application.h"
 #include "ApplicationMessenger.h"
 #include "GUIWindowLoginScreen.h"
-#include "settings/windows/GUIWindowSettingsProfile.h"
+#include "profiles/Profile.h"
+#include "profiles/dialogs/GUIDialogProfileSettings.h"
+#include "profiles/windows/GUIWindowSettingsProfile.h"
 #include "dialogs/GUIDialogContextMenu.h"
-#include "settings/dialogs/GUIDialogProfileSettings.h"
 #include "GUIPassword.h"
 #ifdef HAS_PYTHON
 #include "interfaces/python/XBPython.h"
@@ -36,7 +37,6 @@
 #include "utils/Weather.h"
 #include "network/Network.h"
 #include "addons/Skin.h"
-#include "settings/Profile.h"
 #include "guilib/GUIMessage.h"
 #include "GUIUserMessages.h"
 #include "guilib/GUIWindowManager.h"
